% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{create_plot}
\alias{create_plot}
\title{Plot dot matrix.}
\usage{
create_plot(mat, aln, use_ggplot = FALSE)
}
\arguments{
\item{mat:}{dot matrix.}

\item{use_ggplot:}{logical. If TRUE, plot using ggplot2 package. Otherwise use base-R function.}
}
\value{
A raster plot
}
\description{
Plot dot matrix.
}
