% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{dot_matrix}
\alias{dot_matrix}
\title{Convert alignment to dot matrix.
This functions takes a pairwise alignment and creates a dot matrix
representing the alignment's path.}
\usage{
dot_matrix(a, b, n = 1)
}
\arguments{
\item{a:}{string with DNA/AA sequence.}

\item{b:}{string with DNA/AA sequence.}

\item{n:}{number of ocurrences of alignment in list.}
}
\value{
dot matrix.
}
\description{
Convert alignment to dot matrix.
This functions takes a pairwise alignment and creates a dot matrix
representing the alignment's path.
}
