% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{dotPlotg}
\alias{dotPlotg}
\title{Create dotplot using ggplot, X,Y points with a match will be}
\usage{
dotPlotg(seq1, seq2, wsize = 10, wstep = 1, nmatch = 10)
}
\arguments{
\item{seq1}{String of first sequence}

\item{seq2}{String of second sequence}

\item{wsize}{Window size}

\item{wstep}{Step size for windows (likely should be 1)}

\item{nmatch}{(Number of matches to plot)}
}
\description{
Create dotplot using ggplot, X,Y points with a match will be
}

