% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{generate_dot_matrix}
\alias{generate_dot_matrix}
\title{This functions takes all pairwise alignments and creates a dot matrix
representing the alignment's path.}
\usage{
generate_dot_matrix(input)
}
\arguments{
\item{filepath:}{either a directory contains all alignments (DNA/AA), or a json file contains all alignments of a single raw DNA/AA sequence.}
}
\value{
a single dot matrix.
}
\description{
This functions takes all pairwise alignments and creates a dot matrix
representing the alignment's path.
}
