% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{matrix_to_dataframe}
\alias{matrix_to_dataframe}
\title{Produce a data frame given a dot matrix. The dot matrix can be gnerated by generate_dot_matrix().}
\usage{
matrix_to_dataframe(mat)
}
\arguments{
\item{mat:}{dot matrix.}
}
\value{
a data frame (Within the dataframe, The integeter values in either columns indiciate where the match/gaps occur, the non-integer values indicates where the potential insertion space are.)
}
\description{
Produce a data frame given a dot matrix. The dot matrix can be gnerated by generate_dot_matrix().
}
