% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mkDotPlotMatrix}
\alias{mkDotPlotMatrix}
\title{Generate a dotplot matrix by comparing sequence 1 and sequence 2 using the
given window size and percent match}
\usage{
mkDotPlotMatrix(seq1, seq2, wsize, wstep, nmatch)
}
\arguments{
\item{seq1}{The first sequence}

\item{seq2}{The second sequence}

\item{wsize}{The window of the size to use}

\item{wstep}{the size in chars for the steps of the moving window. Use wstep == wsize for non-overlapping windows.}

\item{nmatch}{if the number of match per window is greater than or equal to nmatch then a dot is produced.}
}
\value{
Returns a matrix
}
\description{
Generate a dotplot matrix by comparing sequence 1 and sequence 2 using the
given window size and percent match
}

