% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{plot_dot_matrix}
\alias{plot_dot_matrix}
\title{Plot a dot matrix.
This function plots a dot matrix as a raster image.}
\usage{
plot_dot_matrix(input, isDir = T)
}
\arguments{
\item{filepath:}{either a directory contains all alignments (DNA/AA), or a json file}

\item{isDir:}{\code{TRUE} or \code{FALSE}. It is true by default that our filepath is a directory. otherwise, it is a json format.}
}
\value{
A raster plot
}
\description{
Plot a dot matrix.
This function plots a dot matrix as a raster image.
}
