% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{read_input}
\alias{read_input}
\title{Read input alignments.
This function reads DNA/AA alignments from a file or directory.}
\usage{
read_input(input_list)
}
\arguments{
\item{file_path:}{list of paths to json/fasta files or path to directory containing fasta alignments.}
}
\value{
data frame with unique alignments and count.
}
\description{
Read input alignments.
This function reads DNA/AA alignments from a file or directory.
}
